Play with Elastic Search
This commit adds logic to first load lenders' data into Elastic
Search and then use it to calculate a long if the user select the
corresponding option in Swagger or the HTTP request.

This is done only for entertainment purposes to see how one can
use the High Level Java client. No real application would read and
load the same file that it's used for CSV calculations to ES and
then query ES to get the same values.